logging-{winston,bunyan}: format logs to work better with Log Viewer and Error Reporting (#1997)
* logging-{winston,bunyan}: format logs to work better with Log Viewer and Error Reporting

Stackdriver Logs Viewer picks up the summary line from the 'message' field
of the jsonPayload or from the textPayload. See:
https://cloud.google.com/logging/docs/view/logs_viewer_v2#expanding

Furthermore, for error messages at severity 'error' and higher, Stackdriver
Error Reporting will pick up error messages if the full stack trace is
included in the textPayload or in the `message` property of the
jsonPaylod. See:
https://cloud.google.com/error-reporting/docs/formatting-error-messages

Fixes: https://github.com/GoogleCloudPlatform/google-cloud-node/issues/1989

* error reporting prefers jsonPayload

* drop devDep on async

* add TODO about serviceContext.service and global resource types

* bunyan: address comments

* winston: address comments

* winston: remove : when appending the stack

* address comments

* one liner

* ensure test records have a level