feat: request-correlating middleware (#63)
This change introduces a mechanism to export middleware – only express
added at this point – others are possible in the future.

The middleware makes it simple to start logging from express apps
without having to deal with boilerplate to initialize bunyan and the
stackdriver stream separately.

Furthermore, the middleware exposes a `log` API on the `request` object
that can do context-correlation. All log entries corresponding to a
specific request will be bundled together.