fix(middleware): use bunyan log.child
It turns out bunyan provides built-in support for Child loggers to
insert scope specific keys into the log metadata. This means that
we don't have to hand-roll an approximation to the bunyan API.

There should be no functional changes for JS users. TS users will see
the types change. Given that this support is clearly marked as
experimental in the docs, we are not going to deem these as semver
major changes.