fix: do not promote record.labels if improper
We promote a bunyan record.labels property to the LogEntry.labels
property. This enables label based filtering to work auto-magically.

However, this can be problematic when the labels are not type correct
as per the logging proto requirements - specifically, the labels values
have to be strings.

Restrict the automatic promotion to happen only in the cases when
know that the labels are well-formatted.

Fixes: https://github.com/googleapis/nodejs-logging-bunyan/issues/265