test: fix error reporting system test race
The system tests for error reporting were still not safe to run
concurrently. The issue is that at the end we were deleting all events
(there is no API to delete events selectively, AFAICT). This meant that
concurrent executions of the system tests against the same project could
clobber anothers logged errors.

This is consistent with the observed timeouts looking for events to
show up.