Update HttpRequest to match @google-cloud/logging
Currently, emitting a log containing an httpRequest that is strongly typed using types.HttpRequest can result in a couple of errors:

* "Empty string" when sending a field for which we don't have information.
* google.logging.type.HttpRequest.latency: object expected

This PR changes the definition to match that defined in https://github.com/googleapis/nodejs-logging/blob/master/src/http-request.ts. In particular, it makes all fields optional, changes latency to the format that matches the proto, and changes a couple of string fields to numbers.